Code
It ended up being simpler than i thought... I just added a while loop to check what you need to multiply with the inverse of the fraction to get the same modulo. The reason I print with an additional modulo is because it sometimes returns a crazy high number with non 1 fractions.